+++ Practice/Problems/UVa/Bitmanip/11173_one_liner.cpp
@@ -1,0 +1,19 @@
+// 11173, Submission: 26179636 verdict: AC
+#include<iostream>
+
+using namespace std;
+
+int find_gc(int k){
+    return k ^ (k>>1);
+}
+
+int main(){
+    int N;
+    cin>>N;
+    while(N--){
+        int n,k;
+        cin>>n>>k;
+        int num = find_gc(k);
+        cout<<num<<endl;
+    }
+}